Fix race conditions in transmitting polar boundaries
A while back I fixed a race condition in the forced B3
cancellation.  Bafflingly, I didn't fix it for the other three
times the pattern occurred, so this addresses that.

This produced incorrect EMF1 values on the pole, which would lead
to polar instability, so hoping it addresses #129.

It also made averaging of U3/T3 unstable, which I know a few people
are still using.

Whether it is the only relevant bug for these issues remains to be
seen.